refactor(server): switch signaling to internal string ids and session-owned voice state
- remove PublicIdService and send database ids directly in commands and events
- make channel-scoped commands resolve hub context from channel_id instead of hub_id
- rework SessionManager for multi-session ownership, takeover, and staged voice joins
- add VoiceSelfStatus and VoiceSelfRevoked signaling for deterministic voice ownership
- remove infra/security validators and move inbound protobuf parsing into ParsedPayload
- add typed EnvLoader getters and wire websocket limits from MAX_CONNECTIONS and MAX_MESSAGE_SIZE
- cache the serialized app pong response and tighten websocket startup/runtime limits
- add server-only docker startup flow and related Caddy/docker config updates

+++ net/transport/websocket/WebSocketTransport.h
@@ -29,7 +29,7 @@ namespace net::transport::websocket {
 
 struct WsLimits {
     size_t max_message_bytes = 256 * 1024;
-    uint8_t max_connections = 255;
+    size_t max_connections = 255;
 };
 
 class TextWSServer : public ITransportServer,
